requirements.txt
beautifulsoup4==4.4.1
dnspython3==1.12.0
selenium==2.52.0
six==1.10.0
tld==0.7.6
